# DOC: integrate.solve_ivp: clarify impact of parameter `vectorized`