--- a/scipy/integrate/_ivp/base.py
+++ b/scipy/integrate/_ivp/base.py
@@ -64,8 +64,8 @@ class OdeSolver:
     Parameters
     ----------
     fun : callable
-        Right-hand side of the system: the time derivative of the state ``y`` at
-        time ``t``. The calling signature is ``fun(t, y)``, where ``t`` is a
+        Right-hand side of the system: the time derivative of the state ``y``
+        at time ``t``. The calling signature is ``fun(t, y)``, where ``t`` is a
         scalar and ``y`` is an ndarray with ``len(y) = len(y0)``. ``fun`` must
         return an array of the same shape as ``y``. See `vectorized` for more
         information.
@@ -85,13 +85,14 @@ class OdeSolver:
         If ``vectorized`` is True, `fun` may be called with ``y`` of shape
         ``(n, k)``, where ``k`` is an integer. In this case, `fun` must behave
         such that ``fun(t, y)[:, i] == fun(t, y[:, i])`` (i.e. each column of
-        the result is the time derivative of the state corresponding with a
-        column of ``y``).
+        the returned array is the time derivative of the state corresponding
+        with a column of ``y``).
 
         Setting ``vectorized=True`` allows for faster finite difference
-        approximation of the Jacobian by methods 'Radau` and 'BDF', but
-        will result in slower execution for other methods and for 'Radau' and
-        'BDF' in some circumstances (e.g. small ``len(y0)``).
+        approximation of the Jacobian by methods 'Radau' and 'BDF', but
+        will result in slower execution for other methods. It can also
+        result in slower overall execution for 'Radau' and 'BDF' in some
+        circumstances (e.g. small ``len(y0)``).
     support_complex : bool, optional
         Whether integration in a complex domain should be supported.
         Generally determined by a derived solver class capabilities.