#### Reference issue
Closes gh-8922
Closes gh-9061
Closes gh-9728

#### What does this implement/fix?
gh-9728 noted that the meaning of `vectorized` was not as explicit as it could be.
gh-9061 noted an apparent contradiction in the documentation of `scipy.integrate.LSODA`:

> The vectorized implementation allows a faster approximation of the Jacobian by finite differences (required for this solver).

> A vectorized implementation offers no advantages for this solver. Default is False.

I dug a bit deeper, and found that a vectorized `fun` is [only used by BDF and Radau](https://github.com/scipy/scipy/search?q=fun_vectorized), so it can only be advantageous for these solvers.

However, the [documentation](https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.solve_ivp.html) currently guarantees that if `vectorized is True`, `fun` will only be called with 2D `y`. Most of the time, the code wants to call `fun` with 1D `y`, though, so `fun` gets wrapped to accept 1D input and return 1D output.

https://github.com/scipy/scipy/blob/8a64c938ddf1ae4c02a08d2c5e38daeb8d061d38/scipy/integrate/_ivp/base.py#L123-L139

The overhead of using this wrapped function is substantial. It slows down solvers other than BDF and Radau considerably, and it can even slow down BDF and Radau overall unless the bottleneck is in the Jacobian calculation.

```python3
from scipy import integrate
def lotkavolterra(t, z):
    a, b, c, d = (1.5, 1, 3, 1)
    x, y = z
    return [a*x - b*x*y, -c*y + d*x*y]

%timeit integrate.solve_ivp(lotkavolterra, [0, 15], [10, 5], method='Radau')
37.6 ms ± 1.49 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

%timeit integrate.solve_ivp(lotkavolterra, [0, 15], [10, 5], method='Radau', vectorized=True)
44.7 ms ± 200 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

This PR rephrases the documentation to:
- be more explicit (e.g. define `n`, address gh-9728)
- clarify when using `vectorized` is beneficial and when it isn't
- suggest that if `vectorized` is True, `fun` must accept _either_ 1D or 2D `y`. If we can rely on this behavior, we can avoid wrapping `fun` and get the best of both worlds. In many cases, this requires no special effort from the user. For instance, `lotkavolterra` above works exactly how we'd want it to for ND `y`.

#### Additional information
What else should we do?

- Raise a warning when `vectorized=True` but `method not in {'Radau', 'BDF'}`
- Check whether `fun` does the right thing with 1D or 2D `y`, wrap it only if it doesn't, and possibly raise a `DeprecationWarning` if it doesn't.