@Patol75 2023-01-30
Thanks for looking into this @mdhaber. I also remember as a first-time user of `solve_ivp` being extremely confused by the meaning of `vectorized`. Your changes add clarity to the purpose of this argument and, in my opinion, are much welcomed!

About the warnings you mention in **Additional information**, were you thinking to have them implemented as part of this PR or another one?