@mdhaber 2023-01-30
> were you thinking to have them implemented as part of this PR or another one?

Let's make them follow-up PRs. It will be easier to review this PR as a doc-only change.

In any case, which of those other things do you think we should do?