@Patol75 2023-02-08
I would tend to say that raising a warning if a method other than `Radau` or `BDF` is used with `vectorized=True` should be sufficient given the updated documentation. I am happy to approve the PR if you want it merged relatively soon.