@mdhaber 2023-02-18
Thanks @Patol75. Yes, if you wouldn't mind reviewing this one, it would encourage a maintainer to take a look. After this is merged, I'll open a PR to raise warnings.