@mdhaber 2023-02-20
Thanks @Patol75. Hopefully addressed everything in [8b9d6ba](https://github.com/scipy/scipy/pull/17836/commits/8b9d6ba190646433c6d9153deba972898739da76).